Move calc_column_width to new module
