+++ conductr_cli/conduct_info.py
@@ -1,4 +1,4 @@
-from conductr_cli import bundle_utils, conduct_url, conduct_logging
+from conductr_cli import bundle_utils, conduct_url, conduct_logging, screen_utils
 import json
 import requests
 
@@ -27,7 +27,7 @@ def info(args):
     data.insert(0, {'id': 'ID', 'name': 'NAME', 'replications': '#REP', 'starting': '#STR', 'executions': '#RUN'})
 
     padding = 2
-    column_widths = dict(calc_column_widths(data), **{'padding': ' ' * padding})
+    column_widths = dict(screen_utils.calc_column_widths(data), **{'padding': ' ' * padding})
     has_error = False
     for row in data:
         has_error |= '!' in row['id']
@@ -40,14 +40,3 @@ def info(args):
 
     if has_error:
         print('There are errors: use `conduct events` or `conduct logs` for further information')
-
-
-def calc_column_widths(data):
-    column_widths = {}
-    for row in data:
-        for column, value in row.items():
-            column_len = len(str(value))
-            width_key = column + '_width'
-            if column_len > column_widths.get(width_key, 0):
-                column_widths[width_key] = column_len
-    return column_widths
